requirements.txt
# Core dependencies
pydantic>=2.5.0
python-dotenv>=1.0.0

# Async & HTTP
aiohttp>=3.9.1

# Image Processing
Pillow>=10.3.0

# Document Parsing
PyMuPDF>=1.23.8
python-docx>=1.1.0
PyYAML>=6.0.1

# CLI
click>=8.1.7

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Optional: Rich console output
rich>=13.7.0
